backend/requirements.txt
Django>=5.2.13
djangorestframework>=3.14
djangorestframework-simplejwt>=2.0
django-cors-headers>=3.15
django-anymail>=8.3
python-dotenv>=1.0
requests>=2.31
PyJWT>=2.8
psycopg2-binary>=2.9
pdfplumber>=0.9.0
python-docx>=0.8.11
scikit-learn>=1.3.0
sentence-transformers>=2.3.2
faiss-cpu>=1.8.0
